requirements.txt
joblib==1.5.1
numpy==2.1.2
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.0.0
scikit-learn==1.7.0
seaborn==0.13.2
streamlit==1.46.1
torch==2.7.1+cu118
torchsummary==1.5.1
torchvision==0.22.1+cu118